# src/public/openapi.yaml
openapi: 3.0.0
info:
  title: users-service
  version: "0.1.0"
paths:
...
  /oauth/token:
    post:
      operationId: OAuth2Token
      tags:
        - Access Control
      description: OAuth2 Token Implementation
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/OAuth2TokenRequest"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2TokenResponse"
        400:
          $ref: "#/components/responses/OAuthErrorBadRequest"
        401:
          $ref: "#/components/responses/OAuthErrorBadRequest"
components:
  schemas:
...
    OAuth2TokenResponse:
      type: object
      additionalProperties: false
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        token_type:
          type: string
          enum:
          - bearer
        expires_in:
          type: number
          example: 300
        refresh_token:
          type: string
          example: asdgniadin23jfaosdfnsoafnalskdnfo2n3afifnk23jnawe
        scope:
          type: string
      required:
        - access_token
        - token_type
        - expires_in
...
    OAuth2TokenRequest:
      type: object
      properties:
        grant_type:
          type: string
          enum:
          - authorization_code
          - password
          - refresh_token
          - client_credentials
        scope:
          type: string
          nullable: true
        refresh_token:
          type: string
          nullable: true
      required:
        - grant_type
    OAuth2ErrorSchema:
      type: object
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
        error_description:
          type: string
        error_uri:
          type: string
      required:
      - error
  responses:
...
    OAuthErrorBadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OAuth2ErrorSchema"
...
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT